preview_shape_factor is a float.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@4584 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/lyxrc.h
@@ -355,7 +355,7 @@ enum LyXRCTags {
 	///
 	bool preview_hashed_labels;
 	///
-	int preview_scale_factor;
+	float preview_scale_factor;
 	
 private:
 	/// Is a bind file already (or currently) read?
